feat: home page support for team workspaces
Adds a workspace header (logo, name, settings gear for editors,
markdown description) on /home when the current workspace is not
personal, replacing the "Welcome, {name}" greeting. Recent workflows
now scope to the current user's runs in a team workspace (mirroring
the run-history "Just mine" default). Saved-workflow card is
redesigned: thumbnail bounded by title+description, access badge
inline with the title, author and change-notes as full-width rows
with a divider. Consolidates the per-card components into a single
workflows.tsx exporting all card variants and the PreviewContent
helper.
