+++ gooey-gui/app/components/HomePage/HomePage.css
@@ -22,6 +22,30 @@
   line-height: 1;
 }
 
+.saved-card-thumb {
+  width: 72px;
+  align-self: stretch;
+  object-fit: cover;
+}
+
+.saved-card-thumb > img,
+.saved-card-thumb > video,
+.saved-card-thumb .gui-img,
+.saved-card-thumb .gui-video {
+  width: 100%;
+  max-width: 100%;
+  height: 100%;
+  max-height: 100%;
+  margin: 0;
+  object-fit: cover;
+  display: block;
+}
+
+.saved-card-thumb-letter {
+  font-size: 32px;
+  line-height: 1;
+}
+
 .starter-thumb-author {
   width: 20px;
   height: 20px;
